requirements.txt
Django==2.1.1
#psycopg2
mysqlclient
Pillow
var_dump
WeasyPrint
openpyxl
django-ckeditor
django-taggit
django-mptt
django-widget-tweaks
django-bootstrap-breadcrumbs
django-extensions
djangorestframework
markdown
django-filter
django-crispy-forms
django-widget-tweaks
django-simple-history
django-currentuser
django-audit-log
pydash
django-multiselectfield
django-user-accounts
django-export-csv
django-autocomplete-light
jsonfield2
qrcode[pil]
django-clone==0.0.2
django-cloneable
